linux: Avoid -Wsign-conversion warning in buffer overflow check

+++ libusb/os/linux_usbfs.c
@@ -2242,7 +2242,7 @@ static int submit_control_transfer(struct usbi_transfer *itransfer)
 	struct usbfs_urb *urb;
 	int r;
 
-	if (transfer->length - LIBUSB_CONTROL_SETUP_SIZE > MAX_CTRL_BUFFER_LENGTH)
+	if (transfer->length - (int)LIBUSB_CONTROL_SETUP_SIZE > MAX_CTRL_BUFFER_LENGTH)
 		return LIBUSB_ERROR_INVALID_PARAM;
 
 	urb = calloc(1, sizeof(*urb));
